Implement compact cron editor
