+++ projects/cron-editor/src/lib/Localizations.ts
@@ -20,6 +20,7 @@ export class Localizations {
     every: 'Every',
     minutes: 'minute(s)',
     onSecond: 'on second',
+    onMinute: 'on minute',
     hoursOnMinute: 'hour(s) on minute',
     andSecond: 'and second',
     daysAt: 'day(s) at',
@@ -34,6 +35,7 @@ export class Localizations {
     cronExpression: 'Cron expression',
     cronExpressionCannotBeNull: 'Cron expression cannot be null',
     invalidCronExpression: 'Invalid cron expression, there must be {expected} segments',
+    couldNotParseExpression: 'Could not parse cron expression',
     lastDay: 'last day',
     lastWeekday: 'last weekday',
     firstWeekday: 'first weekday',
@@ -96,6 +98,7 @@ export class Localizations {
     every: 'Varje',
     minutes: 'minut',
     onSecond: 'på sekund',
+    onMinute: 'på minut',
     hoursOnMinute: 'timmar på minut',
     andSecond: 'och sekund',
     daysAt: 'dagar på',
@@ -110,6 +113,7 @@ export class Localizations {
     cronExpression: 'Cron-uttryck',
     cronExpressionCannotBeNull: 'Cron-uttryck kan inte vara tomt',
     invalidCronExpression: 'Ogiltigt cron-uttryck, det måste innehålla {expected} segment',
+    couldNotParseExpression: 'Lyckades inte tolka cron-uttryck',
     lastDay: 'sista dagen',
     lastWeekday: 'sista arbetsdagen',
     firstWeekday: 'första arbetsdagen',
