+++ projects/cron-editor/src/lib/cron-editor-tabs/cron-editor-tabs.component.html
@@ -433,6 +433,6 @@
   </div>
 </div>
 
-<div class="row" *ngIf="!state.validation.isValid">
-  <div class="col col-md-12"><code>{{state.validation.errorMessage}}</code></div>
+<div class="row" *ngIf="!validation.isValid">
+  <div class="col col-md-12"><code>{{validation.errorMessage}}</code></div>
 </div>
