+++ projects/cron-editor/src/lib/cron-editor.component.css
@@ -17,7 +17,7 @@
   cursor: pointer;
 }
 
-.tab-pane label {
+label {
   display: inline;
   font-weight: inherit;
 }
